requirements.txt
streamlit
pillow
torch
torchvision
transformers
sentence-transformers
google-generativeai
plotly
numpy